self_python/ch_09/inflation_gap_project/requirements.txt
datakart
matplotlib
openpyxl
pandas
requests
seaborn
python-dotenv